+++ src/Products/PloneMeeting/content/organization.py
@@ -91,6 +91,7 @@ class IPMOrganization(IOrganization):
         value_type=schema.Choice(
             vocabulary="Products.PloneMeeting.vocabularies.itemallstates"),
         required=False,
+        default=[],
     )
 
     form.read_permission(item_advice_edit_states='PloneMeeting.manage_internal_organization_fields')
@@ -102,6 +103,7 @@ class IPMOrganization(IOrganization):
         value_type=schema.Choice(
             vocabulary="Products.PloneMeeting.vocabularies.itemallstates"),
         required=False,
+        default=[],
     )
 
     form.read_permission(item_advice_view_states='PloneMeeting.manage_internal_organization_fields')
@@ -113,6 +115,7 @@ class IPMOrganization(IOrganization):
         value_type=schema.Choice(
             vocabulary="Products.PloneMeeting.vocabularies.itemallstates"),
         required=False,
+        default=[],
     )
 
     form.read_permission(keep_access_to_item_when_advice='PloneMeeting.manage_internal_organization_fields')
@@ -156,6 +159,7 @@ class IPMOrganization(IOrganization):
         value_type=schema.Choice(
             vocabulary="collective.contact.plonegroup.organization_services"),
         required=False,
+        default=[],
     )
 
     model.fieldset('app_parameters',
@@ -182,7 +186,7 @@ class PMOrganization(Organization):
     def get_groups_in_charge(self, the_objects=False):
         """Accessor so it can be called in a TAL expression."""
         res = self.groups_in_charge
-        if the_objects:
+        if res and the_objects:
             res = uuidsToObjects(res, ordered=True)
         return res
 
@@ -196,7 +200,7 @@ class PMOrganization(Organization):
 
     def get_item_advice_states(self, cfg=None):
         res = self.item_advice_states
-        if cfg:
+        if res and cfg:
             tmpres = []
             givenCfgId = cfg.getId()
             for elt in res:
@@ -206,11 +210,11 @@ class PMOrganization(Organization):
             # if nothing redefined for given p_cfg in this organization,
             # use value defined on the cfg
             res = tmpres or cfg.getItemAdviceStates()
-        return tuple(res)
+        return res
 
     def get_item_advice_edit_states(self, cfg=None):
         res = self.item_advice_edit_states
-        if cfg:
+        if res and cfg:
             tmpres = []
             givenCfgId = cfg.getId()
             for elt in res:
@@ -220,11 +224,11 @@ class PMOrganization(Organization):
             # if nothing redefined for given p_cfg in this organization,
             # use value defined on the cfg
             res = tmpres or cfg.getItemAdviceEditStates()
-        return tuple(res)
+        return res
 
     def get_item_advice_view_states(self, cfg=None):
         res = self.item_advice_view_states
-        if cfg:
+        if res and cfg:
             tmpres = []
             givenCfgId = cfg.getId()
             for elt in res:
@@ -234,7 +238,7 @@ class PMOrganization(Organization):
             # if nothing redefined for given p_cfg in this organization,
             # use value defined on the cfg
             res = tmpres or cfg.getItemAdviceViewStates()
-        return tuple(res)
+        return res
 
     def get_keep_access_to_item_when_advice(self, cfg=None):
         """ """